permissions

+++ app/Http/Controllers/Admin/UserController.php
@@ -3,16 +3,30 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Models\Hotel;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\View\View;
+use Illuminate\Http\RedirectResponse;
+use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+        $this->middleware('permission:create-user|edit-user|delete-user', ['only' => ['index','show']]);
+        $this->middleware('permission:create-user', ['only' => ['create','store']]);
+        $this->middleware('permission:edit-user', ['only' => ['edit','update']]);
+        $this->middleware('permission:delete-user', ['only' => ['destroy']]);
+    }
+
+
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request): View
     {
         $hotel = $request->session()->get('hotel_id');
         $users = User::paginate(10);
@@ -25,7 +39,8 @@ class UserController extends Controller
     public function create(Request $request)
     {
         $hotel = $request->session()->get('hotel_id');
-        return view('auth.users.form', compact('hotel'));
+        $roles = Role::pluck('name')->all();
+        return view('auth.users.form', compact('hotel', 'roles'));
     }
 
     /**
@@ -41,30 +56,59 @@ class UserController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Request $request, User $user)
+    public function edit(User $user, Request $request): View
     {
         $hotel = $request->session()->get('hotel_id');
-        return view('auth.users.form', compact('user', 'hotel'));
+        // Check Only Super Admin can update his own Profile
+        if ($user->hasRole('Super Admin')){
+            if($user->id != auth()->user()->id){
+                abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSIONS');
+            }
+        }
+
+        return view('auth.users.form', [
+            'user' => $user,
+            'roles' => Role::pluck('name')->all(),
+            'userRoles' => $user->roles->pluck('name')->all(),
+            'hotel' => $hotel
+        ]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, User $user)
+    public function update(Request $request, User $user): RedirectResponse
     {
-        $params = $request->all();
-        $user->update($params);
-        session()->flash('success', 'Пользователь ' . $user->name . ' обновлен');
-        return redirect()->route('users.index');
+        $input = $request->all();
+
+        if(!empty($request->password)){
+            $input['password'] = Hash::make($request->password);
+        }else{
+            $input = $request->except('password');
+        }
+
+        $user->update($input);
+
+        $user->syncRoles($request->roles);
+
+        return redirect()->route('users.index')
+            ->withSuccess('User is updated successfully.');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(User $user)
+    public function destroy(User $user): RedirectResponse
     {
+        // About if user is Super Admin or User ID belongs to Auth User
+        if ($user->hasRole('Super Admin') || $user->id == auth()->user()->id)
+        {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSIONS');
+        }
+
+        $user->syncRoles([]);
         $user->delete();
-        session()->flash('success', 'Пользователь ' . $user->title . ' удален');
-        return redirect()->route('users.index');
+        return redirect()->route('users.index')
+            ->withSuccess('Пользователь ' . $user->title . ' удален');
     }
 }
